Fix search_wikipedia retries
