+++ sdks/opik_optimizer/scripts/adk_agent.py
@@ -4,6 +4,7 @@ from opik_optimizer import (
     OptimizableAgent,
     ChatPrompt,
 )
+from opik_optimizer.utils import search_wikipedia
 
 from opik.integrations.adk import OpikTracer
 
@@ -14,20 +15,6 @@ from google.adk.sessions import InMemorySessionService
 from google.genai import types
 
 from pydantic import BaseModel, Field
-
-# For wikipedia tool:
-import dspy
-
-
-def search_wikipedia(query: str) -> list[str]:
-    """
-    This agent is used to search wikipedia. It can retrieve additional details
-    about a topic.
-    """
-    results = dspy.ColBERTv2(url="http://20.102.90.50:2017/wiki17_abstracts")(
-        query, k=3
-    )
-    return [item["text"] for item in results]
 
 
 # Input schema used by both agents
