+++ sdks/opik_optimizer/scripts/langgraph_agent.py
@@ -2,7 +2,7 @@ from typing import Any, List, Dict
 from typing_extensions import TypedDict
 
 from opik.integrations.langchain import OpikTracer
-from opik import track
+from opik_optimizer.utils import search_wikipedia
 
 from opik_optimizer import (
     OptimizableAgent,
@@ -13,21 +13,6 @@ from langgraph.graph import StateGraph
 from langchain_openai import ChatOpenAI
 from langchain.agents import Tool, create_react_agent, AgentExecutor
 from langchain_core.prompts import PromptTemplate
-
-# Tools:
-import dspy
-
-
-@track(type="tool")
-def search_wikipedia(query: str) -> list[str]:
-    """
-    This agent is used to search wikipedia. It can retrieve additional details
-    about a topic.
-    """
-    results = dspy.ColBERTv2(url="http://20.102.90.50:2017/wiki17_abstracts")(
-        query, k=3
-    )
-    return [item["text"] for item in results]
 
 
 class InputState(TypedDict):
