+++ sdks/opik_optimizer/scripts/litellm_metaprompt_hotpot_example.py
@@ -5,23 +5,10 @@ from opik_optimizer import (
     MetaPromptOptimizer,
 )
 from opik_optimizer.datasets import hotpot_300
+from opik_optimizer.utils import search_wikipedia
 
 from opik.evaluation.metrics import LevenshteinRatio
 from opik.evaluation.metrics.score_result import ScoreResult
-
-# For wikipedia tool:
-import dspy
-
-
-def search_wikipedia(query: str) -> list[str]:
-    """
-    This agent is used to search wikipedia. It can retrieve additional details
-    about a topic.
-    """
-    results = dspy.ColBERTv2(url="http://20.102.90.50:2017/wiki17_abstracts")(
-        query, k=3
-    )
-    return [item["text"] for item in results]
 
 
 dataset = hotpot_300()
