+++ sdks/opik_optimizer/src/opik_optimizer/utils.py
@@ -19,6 +19,7 @@ import json
 import logging
 import random
 import string
+import time
 import urllib.parse
 from types import TracebackType
 
@@ -354,3 +355,32 @@ def python_type_to_json_type(python_type: type) -> str:
         return "array"
     else:
         return "string"  # default fallback
+
+
+@opik.track(type="tool")
+def search_wikipedia(query: str) -> list[str]:
+    """
+    This agent is used to search wikipedia. It can retrieve additional details
+    about a topic.
+    """
+    try:
+        import dspy
+    except ImportError:
+        raise ImportError(
+            "search_wikipedia requires the Python module 'dspy'"
+        ) from None
+
+    results = None
+    for i in range(4):
+        try:
+            results = dspy.ColBERTv2(url="http://20.102.90.50:2017/wiki17_abstracts")(
+                query, k=3
+            )
+        except KeyError:
+            time.sleep(1)
+            continue
+
+    if results:
+        return [item["text"] for item in results]
+    else:
+        raise Exception("too many retries on dspy.ColBERTv2()")
